rf(api): Chat request API refactored to make it re-usable
